+++ packages/atomic/src/components/search/atomic-refine-modal/atomic-refine-modal.tsx
@@ -187,9 +187,11 @@ export class AtomicRefineModal implements InitializableComponent {
   }
 
   private renderFilters() {
-    const isFacetElements = this.bindings.store.getFacetElements().length > 0;
+    const hasFacetElements = this.bindings.store.getFacetElements().length > 0;
+    const hasAutomaticFacets =
+      this.bindings.engine.state.automaticFacetSet?.set !== undefined;
 
-    if (!isFacetElements) {
+    if (!hasFacetElements && !hasAutomaticFacets) {
       return;
     }